x86: Use int32_t for disp and imm in struct x86_insn
Immediate and displacement are always 4 bytes on 32-bit. On 64-bit, immediates
can be 8 bytes in special cases but as we don't support that now, make ->disp
and ->imm "int32_t".

+++ arch/x86/encode.c
@@ -47,8 +47,8 @@ struct x86_insn {
 	uint8_t			opc_ext;
 	uint8_t			mod_rm;
 	uint8_t			sib;
-	int			imm;
-	long			disp;
+	int32_t			imm;
+	int32_t			disp;
 };
 
 static uint8_t x86_register_numbers[] = {
